feat: frontend code overall organization
